address autoloading issues
