+++ lib/cypress.rb
@@ -8,6 +8,7 @@ require_relative '../app/uploaders/patient_archive_uploader'
 require_relative '../app/uploaders/supplement_uploader'
 
 require_relative 'cypress/error_collector'
+require_relative 'cypress/npi_generator'
 require_relative 'cypress/product_status_values'
 require_relative '../app/helpers/test_executions_results_helper'
 
@@ -26,13 +27,13 @@ require_relative '../app/helpers/version_config_helper'
 require_relative '../app/helpers/xml_view_helper'
 
 require_relative 'ext/bundle'
-require_relative 'ext/data_element'
-require_relative 'ext/id'
-require_relative 'ext/individual_result'
-require_relative 'ext/measure'
+require_relative 'ext/qdm/data_element'
+require_relative 'ext/qdm/id'
+require_relative 'ext/cqm/individual_result'
+require_relative 'ext/cqm/measure'
 require_relative 'ext/moped'
-require_relative 'ext/patient'
-require_relative 'ext/provider'
+require_relative 'ext/cqm/patient'
+require_relative 'ext/cqm/provider'
 require_relative 'ext/string'
 require_relative 'ext/value_set'
 require_relative 'ext/zip_file_generator'
@@ -59,7 +60,7 @@ require_relative 'validators/qrda_cat1_validator'
 require_relative 'validators/qrda_cat3_validator'
 require_relative 'validators/qrda_upload_validator'
 
-require_relative 'cypress/highlighting/highlighting'
+require_relative 'cypress/highlighting'
 
 require_relative 'cypress/api_measure_evaluator'
 require_relative 'cypress/clinical_randomizer'
@@ -76,7 +77,6 @@ require_relative 'cypress/expected_results_calculator'
 require_relative 'cypress/mbi_generator'
 require_relative 'cypress/measure_evaluator'
 require_relative 'cypress/name_randomizer'
-require_relative 'cypress/npi_generator'
 require_relative 'cypress/patient_filter'
 require_relative 'cypress/patient_zipper'
 require_relative 'cypress/population_clone_job'
