+++ lib/cypress/api_measure_evaluator.rb
@@ -489,7 +489,7 @@ module Cypress
 
     def do_calculation(product_test, patients, correlation_id)
       measures = product_test.measures
-      calc_job = Cypress::CqmExecutionCalc.new(patients.map(&:qdmPatient), measures, correlation_id,
+      calc_job = Cypress::CQMExecutionCalc.new(patients.map(&:qdmPatient), measures, correlation_id,
                                                effectiveDate: Time.at(product_test.measure_period_start).in_time_zone.to_formatted_s(:number))
       calc_job.execute
     end
