+++ lib/cypress/expected_results_calculator.rb
@@ -8,7 +8,7 @@ module Cypress
     # @param [Array] patients the list of patients that are included in the aggregate results
     # @param [String] correlation_id the id used to associate a group of patients
     # @param [String] effective_date used when generating the query_cache_object for HDS QRDA Cat III export
-    # @param [Hash] options :individual_results are the raw results from CqmExecutionCalc
+    # @param [Hash] options :individual_results are the raw results from CQMExecutionCalc
     def initialize(patients, correlation_id, effective_date)
       @correlation_id = correlation_id
       # Hash of patient_id and their supplemental information
