+++ lib/cypress/highlighting_clause.mustache
