+++ lib/ext/cqm/measure.rb
