+++ lib/ext/cqm/patient.rb
