+++ lib/ext/qdm/id.rb
