+++ lib/validators/calculating_augmented_records.rb
@@ -25,7 +25,7 @@ module Validators
       product_test = ProductTest.find(record.correlation_id)
       return false unless record
 
-      calc_job = Cypress::CqmExecutionCalc.new([record.qdmPatient], product_test.measures, nil,
+      calc_job = Cypress::CQMExecutionCalc.new([record.qdmPatient], product_test.measures, nil,
                                                effectiveDate: Time.at(product_test.measure_period_start).in_time_zone.to_formatted_s(:number))
       results = calc_job.execute(save: false)
       compare_results(results, record, options)
